# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IMessage": {
				"properties": {
					"id": {
						"type": "string"
					},
					"sender": {
						"type": "string"
					},
					"isRead": {
						"type": "boolean"
					},
					"text": {
						"type": "string"
					},
					"isOutgoing": {
						"type": "boolean"
					},
					"dateCreate": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"isRead",
					"text",
					"isOutgoing",
					"dateCreate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IParserBooking": {
				"properties": {
					"Hotel": {
						"type": "string"
					},
					"RoomType": {
						"type": "string"
					},
					"CheckIn": {
						"type": "string"
					},
					"CheckOut": {
						"type": "string"
					},
					"Booked": {
						"type": "string"
					},
					"Voucher": {
						"type": "string"
					},
					"Board": {
						"type": "string"
					},
					"Market": {
						"type": "string"
					},
					"Remark": {
						"type": "string"
					},
					"Status": {
						"type": "string"
					},
					"Comments": {
						"type": "string"
					},
					"Flight_Arr": {
						"type": "string"
					},
					"Flight_Arr_Time": {
						"type": "string"
					},
					"Flight_Dep": {
						"type": "string"
					},
					"Flight_Dep_Time": {
						"type": "string"
					},
					"Names": {
						"items": {
							"properties": {
								"birthDate": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"Hotel",
					"RoomType",
					"CheckIn",
					"CheckOut",
					"Booked",
					"Voucher",
					"Board",
					"Market",
					"Status",
					"Names"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IParserBookingResponse": {
				"properties": {
					"Adults": {
						"type": "number",
						"format": "double"
					},
					"Age1": {
						"type": "number",
						"format": "double"
					},
					"Age2": {
						"type": "number",
						"format": "double"
					},
					"Age3": {
						"type": "number",
						"format": "double"
					},
					"Age4": {
						"type": "number",
						"format": "double"
					},
					"Age5": {
						"type": "number",
						"format": "double"
					},
					"Age6": {
						"type": "number",
						"format": "double"
					},
					"Age7": {
						"type": "number",
						"format": "double"
					},
					"Board": {
						"type": "string"
					},
					"CheckIn": {
						"type": "string"
					},
					"CheckOut": {
						"type": "string"
					},
					"Children": {
						"type": "number",
						"format": "double"
					},
					"ConfirmationNo": {
						"type": "string"
					},
					"Hotel": {
						"type": "string"
					},
					"Name1": {
						"type": "string"
					},
					"Name2": {
						"type": "string"
					},
					"Name3": {
						"type": "string"
					},
					"Name4": {
						"type": "string"
					},
					"Name5": {
						"type": "string"
					},
					"Name6": {
						"type": "string"
					},
					"Name7": {
						"type": "string"
					},
					"PriceAmount": {
						"type": "number",
						"format": "double"
					},
					"PriceCurrency": {
						"type": "string"
					},
					"ResponseText": {
						"type": "string"
					},
					"ResvID": {
						"type": "number",
						"format": "double"
					},
					"RoomType": {
						"type": "string"
					},
					"Vocher": {
						"type": "string"
					},
					"isCancelled": {
						"type": "string"
					}
				},
				"required": [
					"Adults",
					"Age1",
					"Board",
					"CheckIn",
					"CheckOut",
					"Children",
					"ConfirmationNo",
					"Hotel",
					"Name1",
					"PriceAmount",
					"PriceCurrency",
					"ResponseText",
					"ResvID",
					"RoomType",
					"Vocher",
					"isCancelled"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITourist": {
				"properties": {
					"name": {
						"type": "string"
					},
					"birthDate": {
						"type": "string"
					},
					"sex": {
						"type": "string"
					},
					"hotelServiceId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IHotelServiceBooking": {
				"properties": {
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"serviceName": {
						"type": "string"
					},
					"bookingCode": {
						"type": "string"
					},
					"hotelId": {
						"type": "number",
						"format": "double"
					},
					"hotel": {
						"type": "string"
					},
					"pansionId": {
						"type": "number",
						"format": "double"
					},
					"pansion": {
						"type": "string"
					},
					"roomTypeId": {
						"type": "number",
						"format": "double"
					},
					"roomType": {
						"type": "string"
					},
					"roomMapCode": {
						"type": "string"
					},
					"roomAccommodationId": {
						"type": "number",
						"format": "double"
					},
					"roomAccommodation": {
						"type": "string"
					},
					"roomCategoryId": {
						"type": "number",
						"format": "double"
					},
					"roomCategory": {
						"type": "string"
					},
					"confirmationNumber": {
						"type": "string"
					},
					"checkIn": {
						"type": "string"
					},
					"checkOut": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"tourists": {
						"items": {
							"$ref": "#/components/schemas/ITourist"
						},
						"type": "array"
					},
					"priceRemark": {
						"type": "string"
					}
				},
				"required": [
					"serviceId",
					"serviceName",
					"bookingCode",
					"hotelId",
					"hotel",
					"pansionId",
					"pansion",
					"roomTypeId",
					"roomType",
					"roomMapCode",
					"roomAccommodationId",
					"roomAccommodation",
					"roomCategoryId",
					"roomCategory",
					"confirmationNumber",
					"checkIn",
					"checkOut",
					"status",
					"note",
					"tourists"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFlight": {
				"properties": {
					"flightArr": {
						"type": "string"
					},
					"flightDep": {
						"type": "string"
					}
				},
				"required": [
					"flightArr",
					"flightDep"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBooking": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"bookingName": {
						"type": "string"
					},
					"bookingId": {
						"type": "number",
						"format": "double"
					},
					"action": {
						"type": "string"
					},
					"creationDate": {
						"type": "string"
					},
					"cancelDate": {
						"type": "string"
					},
					"changeDate": {
						"type": "string"
					},
					"marketId": {
						"type": "number",
						"format": "double"
					},
					"marketName": {
						"type": "string"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/IMessage"
						},
						"type": "array"
					},
					"parser": {
						"properties": {
							"manual": {
								"properties": {},
								"additionalProperties": {
									"properties": {
										"message": {
											"type": "string"
										},
										"booking": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"booking"
									],
									"type": "object"
								},
								"type": "object"
							},
							"response": {
								"$ref": "#/components/schemas/IParserBookingResponse"
							},
							"send": {
								"$ref": "#/components/schemas/IParserBooking"
							}
						},
						"required": [
							"manual",
							"response",
							"send"
						],
						"type": "object"
					},
					"hotelService": {
						"$ref": "#/components/schemas/IHotelServiceBooking"
					},
					"hotelServices": {
						"items": {
							"$ref": "#/components/schemas/IHotelServiceBooking"
						},
						"type": "array"
					},
					"flightInfo": {
						"$ref": "#/components/schemas/IFlight"
					},
					"dateInputed": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"bookingName",
					"bookingId",
					"action",
					"marketId",
					"marketName",
					"messages",
					"hotelServices"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IHotel": {
				"properties": {
					"_id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"resort": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"regionId": {
						"type": "number",
						"format": "double"
					},
					"resortId": {
						"type": "number",
						"format": "double"
					},
					"parserCode": {
						"type": "number",
						"format": "double"
					},
					"parserName": {
						"type": "string"
					},
					"parserHotelServer": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"resort",
					"code",
					"category",
					"regionId",
					"resortId",
					"parserHotelServer"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt-passport": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "hotels-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Vassil Shterev ",
			"email": "vsterev@gmail.com"
		}
	},
	"paths": {
...
		"/hotels/mapped": {
			"get": {
				"operationId": "GetAllMapped",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"hotels"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": []
			}
		},
		"/hotels": {
			"get": {
				"operationId": "FindHotelByname",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IHotel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"hotels"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "hotelName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "HotelMap",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IHotel"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"hotels"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"hotelId": {
										"type": "number",
										"format": "double"
									},
									"parserCode": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"hotelId",
									"parserCode"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bookings": {
			"get": {
				"operationId": "GetAllBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IBooking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"bookings"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isByDate",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"yes",
								"no"
							]
						}
					},
					{
						"in": "query",
						"name": "booking",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateTo",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bookings/length": {
			"get": {
				"operationId": "GetLength",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"bookings"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isByDate",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"yes",
								"no"
							]
						}
					},
					{
						"in": "query",
						"name": "booking",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateTo",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateFrom",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bookings/confirm": {
			"post": {
				"operationId": "ConfirmBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"confirmationNumber": {
										"type": "string"
									},
									"booking": {
										"type": "string"
									}
								},
								"required": [
									"message",
									"confirmationNumber",
									"booking"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bookings/deny": {
			"post": {
				"operationId": "DenyBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"booking": {
										"type": "string"
									}
								},
								"required": [
									"message",
									"booking"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}